+++ src/main/java/com/getcode/service/project/ProjectService.java
@@ -362,27 +362,6 @@ public class ProjectService {
     }
 
 
-    @Transactional(readOnly = true)
-    public List<ProjectInfoResponseDto> getMyProject(String memberEmail) {
-
-        Member member = memberRepository.findByEmail(memberEmail).orElseThrow(NotFoundMemberException::new);
-
-        List<Project> projects = member.getProjects();
-
-        List<ProjectInfoResponseDto> myProjectRes = new ArrayList<>();
-
-        projects.forEach(project -> myProjectRes.add(new ProjectInfoResponseDto(project)));
-
-        return myProjectRes;
-
-    }
-
-    public List<ProjectInfoResponseDto> getMyWishProject(String memberEmail) {
-
-        Member member = memberRepository.findByEmail(memberEmail).orElseThrow(NotFoundMemberException::new);
-
-        member.getWishProject();
 
 
-    }
 }
